requirements.txt
pandas==1.1.2
boto3==1.15.3
numpy==1.19.2
streamlit==0.68.0
Pillow==8.0.0
yfinance==0.1.55
